// frontend/src/app/models/product.model.ts
export class ProductModel {
    id:string;
    categoryId:string;
    name:string;
    description:string;
    price:number;
    isActiveOnMonday: boolean;
    isActiveOnTuesday: boolean;
    isActiveOnWednesday: boolean;
    isActiveOnThursday: boolean;
    isActiveOnFriday: boolean;
    isActiveOnSaturday: boolean;
    isActiveOnSunday: boolean;
}